% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nivopie.R
\name{nivopie-shiny}
\alias{nivopie-shiny}
\alias{nivopieOutput}
\alias{renderNivopie}
\alias{nivopie_html}
\title{Shiny bindings for nivopie}
\usage{
nivopieOutput(outputId, width = "100\%\%", height = "400px")

renderNivopie(expr, env = parent.frame(), quoted = FALSE)

nivopie_html(id, style, class, ...)
}
\arguments{
\item{outputId}{output variable to read from}

\item{expr}{An expression that generates a nivopie}

\item{env}{The environment in which to evaluate \code{expr}.}

\item{quoted}{Is \code{expr} a quoted expression (with \code{quote()})? This
is useful if you want to save an expression in a variable.}
}
\description{
Output and render functions for using nivopie within Shiny
applications and interactive Rmd documents.
}
