% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nivopie.R
\name{nivopie}
\alias{nivopie}
\title{nivopie}
\usage{
nivopie(data = NULL, ..., width = NULL, height = NULL,
  elementId = NULL)
}
\description{
This is an htmlwidget that draw an interactive pie chart
}
\examples{
id=c("A", "B", "C")
value=c(12,13,45)
df2=data.frame(id,value)
nivopie(df2)
}
